implement support for building the image on BSD
Some BSD-based operating systems support Linux Docker images by
implementing a `docker` command that has a backend powered by
virtualization and an internal virtualized Linux kernel.

The result is that all code running inside the container is
only exposed to a Linux kernel and is unaware of BSD,
but all code running outside the container is directly exposed
to the BSD-based environment

BSD-based operating systems used to develop this:

- MacOS
- FreeBSD
- OpenBSD

It's often hard to find where exactly the `gtar`, `gar`, and `bash`-version-4+ commands are on BSD-based operating systems,
so if you see errors related to those, try to find and install the dependencies for your OS and then add them to `$PATH`.